saving intermediate results of notebooks that fail
Meant to be used with notebooks output using `nbexplode`, this code
allows for running `nbrun -s notebook.ipynb` and save the portions of
the __save__ list which are defined into a notebook_errors.npz, and the
addition of a new cell which can load this npz.

Undesireable side effect of this is that you have to re-generate the
notebook in order to get the inserted cell to be removed, or remove it
by hand.